api-tests/requirements.txt
allure-behave==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
behave==1.2.6
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
parse==1.20.2
parse-type==0.6.2
pluggy==1.5.0
PyHamcrest==2.1.0
requests==2.32.3
six==1.16.0
urllib3==2.2.2
